Check for archive file before zipping it

+++ DLAJobStreamFunctions.sh
@@ -109,8 +109,12 @@ createArchive ()
                 tar -cvf ${ARC_FILE} ${FILE} >/dev/null 2>&1
             fi
         done
-        rm -f ${ARC_FILE}.gz
-        gzip ${ARC_FILE}
+
+        if [ -f ${ARC_FILE} ]
+        then
+            rm -f ${ARC_FILE}.gz
+            gzip ${ARC_FILE}
+        fi
     fi
 }
